chore: rename helper to be more generic for admin sorting links
