+++ app/views/custom/admin/proposals/_search_form.html.erb
@@ -61,16 +61,16 @@
 
   <div class="small-12 medium-3 column">
     <%= select_tag :tag_name,
-                   options_for_select(proposals_tags_select_options(::Proposal.where(id: @proposals.pluck(:id)), "tags"), params[:tag_name]),
+                   options_for_select(admin_tags_select_options(::Proposal.where(id: @proposals.pluck(:id)), "tags"), params[:tag_name]),
                    { prompt: t("admin.budget_investments.index.tags_filter_all") } %>
   </div>
 
   <div class="small-12 medium-3 column">
-    <%= select_tag :goal, proposals_goal_options,
+    <%= select_tag :goal, admin_goal_options(params[:goal]),
                       include_blank: t("shared.advanced_search.goal_blank") %>
   </div>
   <div class="small-12 medium-3 column">
-    <%= select_tag :target, proposals_target_options,
+    <%= select_tag :target, admin_target_options(params[:target]),
                       include_blank: t("shared.advanced_search.target_blank") %>
   </div>
