+++ app/views/custom/admin/proposals/index.html.erb
@@ -4,11 +4,11 @@
 
 <h2><%= t("admin.proposals.index.title") %></h2>
 
-<%= render "search_form" %>
-
 <%= render "admin/proposals/subnav", i18n_namespace: "admin.proposals.index", active: "proposals" %>
 
 <% if @proposals.any? %>
+  <%= render "search_form" %>
+
   <%= link_to t("admin.budget_investments.index.download_current_selection"),
             admin_proposals_path(csv_params),
             class: "float-right small clear" %>
@@ -17,12 +17,12 @@
   <table>
     <thead>
       <tr>
-        <th class="text-center"><%= link_to_proposals_sorted_by :id %></th>
-        <th data-field="title"><%= link_to_proposals_sorted_by :title %></th>
+        <th class="text-center"><%= link_to_sorted_admin_resource "proposals", :id %></th>
+        <th data-field="title"><%= link_to_sorted_admin_resource "proposals", :title %></th>
         <th><%= t("admin.proposals.index.author") %></th>
         <th><%= t("admin.proposals.index.milestones") %></th>
-        <th data-field="supports" ><%= link_to_proposals_sorted_by :supports %></th>
-        <th data-field="created_at" ><%= link_to_proposals_sorted_by :created_at %></th>
+        <th data-field="supports" ><%= link_to_sorted_admin_resource "proposals", :supports %></th>
+        <th data-field="created_at" ><%= link_to_sorted_admin_resource "proposals", :created_at %></th>
         <th><%= t("admin.proposals.index.sdg") %></th>
         <th><%= t("admin.proposals.index.target") %></th>
         <th><%= t("admin.proposals.index.tags") %></th>
